Header Finished
I finished the header and got my header fonts loaded in with the
fontface rule. Unfortunately I'm using adobe typekit for my bodyfont,
and really don't have time to set up my own webserver to test this. I
really just need to keep pressing on. Next up, going to be working on
the body section. and getting it all formatted. Following that, It's
going to be time to pull together more images and verbiage for each
portfolio item.