Gallery transitions working
Finally got the gallery lightbox transitions working. I tried for about
2 hours using a tutorial that was overly complicated and just wasn't
working. I finally just used my own brain and found a much simpler
solution that actually works with my framework. Next up, I need to see
if creating a javascript lightbox effect is worth doing. If not, I need
to start making the individual portfolio pages layout. Once that's done,
it'll be time for actual content to fill this thing up. After that, it
needs to be QA checked, then a mobile interface needs to be made.