Gallery Working, About section created
Progress is moving along swimmingly. I styled the gallery to have circle
images and made them spaced properly. The about section I created right
under the landing area and added a black BG to it so there is some
immediate differentiation there. There isn't much left to do with the
home page. Adding another section for the Illustration gallery,
formatting my typography, text, inserting a contact section at the
bottom, fixing up the top nav bar, then adding in content.

The big monster is going to be formatting this all for mobile. Ugh. I'm
really not looking forward to that.