Nearly Completed Artifact Page, worked on typography
I've nearly completed the template for what an individual page would
look like for a portfolio piece. It's complete with a CSS only masonry
layout, which was surprisingly easy to get up and running. I've also
worked with the typography quite a bit, and edited a ton of sizes, added
my own custom dropcap class, and am using a placeholder font, Arial
Black, and Arial. I'm going to replace them all as soon as I can get my
hands on Helvetica. This project is really coming together. I got a huge
load completed this evening.