finished artifact templates edited nav menu completed 2 artifacts
I did a lot since my last update. First off, I completed the individual
artifact page by adding a "skills and programs" section, along with
relevant icons. I added functionality to the masonry image gallery where
it turns green when an item is moused over and links off site to the
corresponding image. Lastly, I cleaned up and simplified the navbar
greatly. It's a lot less cluttered now.

Tomorrow the front page needs to be beautified. In addition, more
artifacts will need to be added. I think at this point the site is
working well, and now we need to fill it with lots of stuff.