Cross broswer compatibility complete, typography ready, uploaded!
It's in god's hands now. I've basically finished the website for now. It
just needs my logo up at the top left area, as well as media queries for
mobile functionality. That's going to be the tough part. For now, I've
uploaded the entire thing, and it should be ready by tomorrow for a
presentation. Good night.