Finished responsive design
Or the bulk of it at least... The main mobile and desktop viewports all
have been tweaked and redesigned to look top notch. I'm really proud of
the work I've done here. It's been quite the adventure. The last of it
that needs to be done I'd say is imac displays, and large scale phones.
Even so, those designs don't exactly break with what I currently have,
they just look a little unoptimized.

Small tweaks and this website's done.