Correcting some details
I changed the UI a little bit, moving my bio area from the top to near
the bottom. This left me with the task of being forced to change the
gallery section to black, and the about section to white. Along the way
I had to edit a lot of spacing details and add some classes to fix
things up. No doubt this has created mobile complications, so I'll have
to go back and fix those up tomorrow

Aside from that, I've adjusted the spacing on my navbar. More changes to
come unfortunately. Gotta make sure this thing is as polished as
possible before moving on to any new projects.