Square gallery jquery plugin, css edits
working on integrating a Jquery plugin that resizes and makes my images
into squares. The trouble with this has been it's integration, all of my
current CSS has been messing with the plugins functionality, but at the
same time the plugin lacks some functionality that my CSS has afforded
me. So it's been an awkward balancing act that I've only just found a
solution for. All I need to do next, is clean up my design and make it
respond to different layouts.